[#510] Refactor ClusterFeatures to OT::Cluster::Features
